move tabs from styled to aria
