+++ app/gui/src/dashboard/layouts/AssetPanel/components/AssetPanelTabs.tsx
@@ -4,14 +4,18 @@ import { memo, useCallback, useRef, type ReactNode } from 'react'
 import { AnimatePresence, motion } from 'framer-motion'
 
 import { AnimatedBackground } from '#/components/AnimatedBackground'
+import {
+  Tab,
+  TabList,
+  TabPanel,
+  Tabs,
+  type TabListProps,
+  type TabPanelProps,
+  type TabPanelRenderProps,
+  type TabProps,
+  type TabsProps,
+} from '#/components/aria'
 import { useVisualTooltip } from '#/components/AriaComponents'
-import type {
-  TabListProps,
-  TabPanelProps,
-  TabPanelRenderProps,
-  TabProps,
-} from '#/components/styled/Tabs'
-import { Tab, TabList, TabPanel, Tabs, type TabsProps } from '#/components/styled/Tabs'
 import { Suspense } from '#/components/Suspense'
 import SvgMask from '#/components/SvgMask'
